# application/services/custom_services/severity/src/swagger/swagger.yaml
---
openapi: 3.0.0
info:
   version: 1.3.0
   title: meanticketapp76 open api specification
   description: severity API
   termsOfService: http://api_url/terms/
   contact:
      name: ContactName
      email: hello@wolox.co
      url: https://www.wolox.com.ar/
   license:
      name: Apache 2.0
      url: https://www.apache.org/licenses/LICENSE-2.0.html
servers:
   - url: http://localhost:8015
     description: severity
tags:
- name: severity
paths:
   "/severity/{id}":
      delete:
         tags:
         -  severity
         description: GpDelete
         parameters:
         -  in: "path"
            description: GpDelete
            name: id
            required: true
            schema:
               type: "string"
         responses:
            200:
               description: Success
            400:
               description: Bad Request
            404:
               description: Not Found
            500:
               description: Internal Server Error
      get:
         tags:
         -  severity
         description: GpGetNounById
         parameters:
         -  in: "path"
            description: GpGetNounById
            name: id
            required: true
            schema:
               type: "string"
         responses:
            200:
               description: Success
            400:
               description: Bad Request
            404:
               description: Not Found
            500:
               description: Internal Server Error
   "/severity/get/search":
      get:
         tags:
         -  severity
         description: GpSearch
         responses:
            200:
               description: Success
            400:
               description: Bad Request
            404:
               description: Not Found
            500:
               description: Internal Server Error
   "/severity/get/update":
      put:
         tags:
         -  severity
         description: GpSearchForUpdate
         requestBody:
            content:
               application/json:
                  schema:
                     $ref: "#/components/schemas/severity"
         responses:
            200:
               description: Success
            400:
               description: Bad Request
            404:
               description: Not Found
            500:
               description: Internal Server Error
   "/severity":
      put:
         tags:
         -  severity
         description: GpUpdate
         requestBody:
            content:
               application/json:
                  schema:
                     $ref: "#/components/schemas/severity"
         responses:
            200:
               description: Success
            400:
               description: Bad Request
            404:
               description: Not Found
            500:
               description: Internal Server Error
      get:
         tags:
         -  severity
         description: GpGetAllValues
         responses:
            200:
               description: Success
            400:
               description: Bad Request
            404:
               description: Not Found
            500:
               description: Internal Server Error
      post:
         tags:
         -  severity
         description: GpCreate
         requestBody:
            content:
               application/json:
                  schema:
                     $ref: "#/components/schemas/severity"
         responses:
            200:
               description: Success
            400:
               description: Bad Request
            404:
               description: Not Found
            500:
               description: Internal Server Error
   "/severity/userid/created_by":
      get:
         tags:
         -  severity
         description: GpGetNounCreatedBy
         responses:
            200:
               description: Success
            400:
               description: Bad Request
            404:
               description: Not Found
            500:
               description: Internal Server Error
components:
   schemas:
      severity:
         type: object
         properties:
            name:
               type: "string"
               description: Description
            description:
               type: "string"
               description: Description